GPencil: New Length modifier
Extend or shrink the strokes length. This is very useful for the future Line Art module, giving to the strokes a more "organic" feeling.

This modifier was initially created by @NicksBest but never was merged into `master` branch.

Using this original code, I have adapted it to the new UI and some minor changes in the code.

Differential Revision: https://developer.blender.org/D8264