GPencil: Asset Manager support (LITE)
This patch replaces D13193 and remove any transform tool in the import. Now the import must be done in Edit mode and
all transformations must be done with the standard transform tools of Blender.

A new operator has been added to create an asset base on one of the next option:

* Active Layer.
* All Layers.
* All Layers Separated
* Active Frame (Active Layer).
* Active Frame (All Layers).
* Selected Frames.
* Selected Strokes.

The menu uses `Create Asset`instead of `Mark Asset` because the asset is created as new datablock and not reuse existing data.

To import asset, set the grease pencil object where you want import in Edit mode.

Also new asset preview for Grease Pencil has been added.

For testing, be sure to enable `Experimental Feature`and `Developers Extra`

Differential Revision: https://developer.blender.org/D15271